Fix chat head buttons not showing
We need to use `converse-icon` now that the fontawesome fonts aren't
included
