+++ src/plugins/chatview/utils.js
@@ -1,5 +1,5 @@
 import { __ } from 'i18n';
-import { _converse, api } from "@converse/headless/core";
+import { _converse, api } from '@converse/headless/core';
 import { html } from 'lit';
 
 export function clearHistory (jid) {
@@ -8,35 +8,12 @@ export function clearHistory (jid) {
     }
 }
 
-export async function getHeadingDropdownItem (promise_or_data) {
-    const data = await promise_or_data;
-    return data ? html`
-        <a href="#" class="dropdown-item ${data.a_class}" @click=${data.handler} title="${data.i18n_title}">
-            <converse-icon size="1em" color="var(--text-color-lighten-15-percent)" class="fa ${data.icon_class}"></converse-icon>
-            ${data.i18n_text}
-        </a>
-    ` : '';
-}
-
-export async function getHeadingStandaloneButton (promise_or_data) {
-    const data = await promise_or_data;
-    return html`
-        <a
-            href="#"
-            class="chatbox-btn ${data.a_class} fa ${data.icon_class}"
-            @click=${data.handler}
-            title="${data.i18n_title}"
-        ></a>
-    `;
-}
-
 export async function clearMessages (chat) {
     const result = await api.confirm(__('Are you sure you want to clear the messages from this conversation?'));
     if (result) {
         await chat.clearMessages();
     }
 }
-
 
 export async function parseMessageForCommands (chat, text) {
     const match = text.replace(/^\s*/, '').match(/^\/(.*)\s*$/);
@@ -56,7 +33,7 @@ export async function parseMessageForCommands (chat, text) {
          *      return handled;
          *  }
          */
-        handled = await api.hook('parseMessageForCommands', {model: chat, text}, handled);
+        handled = await api.hook('parseMessageForCommands', { model: chat, text }, handled);
         if (handled) {
             return true;
         }
