+++ src/plugins/controlbox/styles/_controlbox.scss
@@ -415,26 +415,6 @@
                     }
                 }
 
-                .controlbox-head {
-                    display: flex;
-                    flex-direction: row-reverse;
-                    flex-wrap: nowrap;
-                    justify-content: space-between;
-                    min-height: 0;
-
-                    .brand-heading {
-                        color: var(--controlbox-text-color);
-                        font-size: 2em;
-                    }
-                    .chatbox-btn {
-                        color: var(--controlbox-head-color);
-                        margin: 0;
-                        converse-icon {
-                            color: var(--controlbox-head-color);
-                        }
-                    }
-                }
-
                 #converse-register, #converse-login {
                     @include make-col(12);
                     padding-bottom: 0;
