+++ src/plugins/headlines-view/styles/headlines-head.scss
@@ -1,0 +1,53 @@
+.conversejs {
+    .chatbox {
+        &.headlines {
+            converse-headlines-heading {
+                &.chat-head {
+                    .chatbox-title--no-desc {
+                        padding: 0.75rem 1rem;
+                    }
+
+                    &.chat-head-chatbox {
+                        background-color: var(--headline-head-bg-color);
+                        border-bottom: var(--headline-head-border-bottom);
+                    }
+                    background-color: var(--headline-head-bg-color);
+
+                    .chatbox-title__text {
+                        color: var(--headline-head-text-color) !important;
+                    }
+
+                    converse-dropdown {
+                        .dropdown-menu {
+                            converse-icon {
+                                svg {
+                                    fill: var(--headlines-color);
+                                }
+                            }
+                        }
+                    }
+
+                    .chatbox-btn {
+                        converse-icon {
+                            svg {
+                                fill: var(--headline-head-fg-color);
+                            }
+                        }
+                    }
+                }
+            }
+
+            converse-chats {
+                &.converse-fullscreen {
+                    .chatbox.headlines {
+                        .chat-head {
+                            &.chat-head-chatbox {
+                                background-color: var(--headline-head-bg-color);
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
