+++ src/plugins/headlines-view/styles/headlines.scss
@@ -1,36 +1,6 @@
 .conversejs {
     .chatbox {
-        converse-headlines-heading {
-            &.chat-head {
-                .chatbox-title__text {
-                    color: var(--headline-head-text-color) !important;
-                    background-color: var(--headline-head-bg-color);
-                }
-
-                a, a:visited, a:hover, a:not([href]):not([tabindex]) {
-                    &.chatbox-btn {
-                        &.fa,
-                        &.fas,
-                        &.far {
-                            color: var(--headline-head-text-color);
-                            &.button-on:before {
-                                padding: 0.2em;
-                                background-color: var(--headline-head-text-color);
-                                color: var(--headline-head-bg-color);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
         &.headlines {
-            .chat-head {
-                &.chat-head-chatbox {
-                    background-color: var(--headline-head-bg-color);
-                    border-bottom: var(--headline-head-border-bottom);
-                }
-            }
             .chat-body {
                 background-color: var(--background);
                 .chat-message {
@@ -71,11 +41,6 @@
             .chatbox.headlines {
                 .box-flyout {
                     background-color: var(--headline-head-text-color);
-                }
-                .chat-head {
-                    &.chat-head-chatbox {
-                        background-color: var(--headline-head-text-color);
-                    }
                 }
                 .flyout {
                     border-color: var(--headline-head-text-color);
