+++ src/plugins/muc-views/styles/muc-forms.scss
@@ -24,7 +24,7 @@ converse-muc-config-form {
                     margin: 0 0.5em;
                 }
                 .button-primary {
-                    background-color: var(--chatroom-head-button-color);
+                    background-color: var(--chatroom-head-fg-color);
                 }
             }
