+++ src/plugins/muc-views/templates/muc-head.js
@@ -2,7 +2,7 @@ import 'shared/components/dropdown.js';
 import 'shared/components/rich-text.js';
 import { __ } from 'i18n';
 import { _converse, api } from "@converse/headless/core.js";
-import { getHeadingDropdownItem, getHeadingStandaloneButton } from 'plugins/chatview/utils.js';
+import { getStandaloneButtons, getDropdownButtons } from 'shared/chat/utils.js';
 import { html } from "lit";
 import { until } from 'lit/directives/until.js';
 
@@ -10,6 +10,7 @@ import { until } from 'lit/directives/until.js';
 export default (el) => {
     const o = el.model.toJSON();
     const subject_hidden = el.user_settings?.get('mucs_with_hidden_subject', [])?.includes(el.model.get('jid'));
+    const heading_buttons_promise = el.getHeadingButtons(subject_hidden);
     const i18n_hide_topic = __('Hide the groupchat topic');
     const i18n_bookmarked = __('This groupchat is bookmarked');
     const subject = o.subject ? o.subject.text : '';
@@ -38,12 +39,8 @@ export default (el) => {
                 </div>
             </div>
             <div class="chatbox-title__buttons row no-gutters">
-                ${ until(el.getHeadingButtons(subject_hidden).then((buttons) => {
-                    const dropdown_btns = buttons.filter(b => !b.standalone).map(b => getHeadingDropdownItem(b));
-                    return html`
-                        ${ buttons.filter(b => b.standalone).reverse().map(b => until(getHeadingStandaloneButton(b), '')) }
-                        ${ dropdown_btns.length ? html`<converse-dropdown class="dropleft" color="var(--chatroom-head-color)" .items=${dropdown_btns}></converse-dropdown>` : '' }`
-                }), '') }
+                ${ until(getStandaloneButtons(heading_buttons_promise), '') }
+                ${ until(getDropdownButtons(heading_buttons_promise), '') }
             </div>
         </div>
         ${ show_subject ? html`<p class="chat-head__desc" title="${i18n_hide_topic}">
