+++ src/shared/styles/themes/concord.scss
@@ -25,7 +25,7 @@
     --chatroom-correcting-color: #FFFFC0;
     --chatroom-head-bg-color: white;
     --chatroom-head-border-bottom: 1px solid #EEE;
-    --chatroom-head-button-color: #999;
+    --chatroom-head-fg-color: #999;
     --chatroom-head-color: #7E7E7E;
     --chatroom-head-description-border-left: 1px solid #DDD;
     --chatroom-head-description-color: black;
