+++ src/shared/tests/tests.css
@@ -1,3 +1,3 @@
-body {
+body.converse-overlayed {
     overflow: auto !important;
 }
